Add debug support for nested scopes
This gets DebugScope and DebugDeclare in the correct order for
nested scopes.

Added Scope and Declare operations to AST, generated only for debug
info generation. Scope sequence nodes define a lexical scope. Declare
defines the point at which a local variable is declared and forces the
definition of the variable in the correct scope.

This commit also generates a DebugScope at the top of every basic block
as is required by the spec.